Initialize logger with prom config from flags
This addresses two problems.

First it looks like the logger is initialized with an empty config, then
a different config is passed into kingpin to be filled up from the
command line arguments.

This change re-initializes the logger with the prom config parsed by
kingpin so that `--log.level` and `--log.format` populate the promlog
instance as expected.

The second issue is that there are 2 go-kit logging libraries.
The smaller `go-kit/log` and `go-kit/kit/log`.

`promlog` is using the smaller `go-kit/log` so `pgpool2_exporter` has
been updated to use the same. Otherwise we were using different
instances of the level types which means the following type assertion
would fail inside the filtered logger since the level is from the other
package:

```go
if v, ok := keyvals[i].(*levelValue); ok {
```